32 bit floating point positioning
Hoping for higher cache hits due to smaller sizes

+++ wasm/src/bindings/camera.rs
@@ -11,45 +11,45 @@ pub static mut CAMERA: objects::camera::Camera = objects::camera::Camera {
     facing: ZEROED_FACING,
 };
 
-const CAMERA_SENSITIVITY: f64 = 10.0;
+const CAMERA_SENSITIVITY: f32 = 10.0;
 
 #[wasm_bindgen]
-pub fn camera_backwards(value: Option<f64>) {
+pub fn camera_backwards(value: Option<f32>) {
     unsafe {
         CAMERA.pos.z -= value.unwrap_or(CAMERA_SENSITIVITY);
     }
 }
 
 #[wasm_bindgen]
-pub fn camera_forwards(value: Option<f64>) {
+pub fn camera_forwards(value: Option<f32>) {
     unsafe {
         CAMERA.pos.z += value.unwrap_or(CAMERA_SENSITIVITY);
     }
 }
 
 #[wasm_bindgen]
-pub fn camera_left(value: Option<f64>) {
+pub fn camera_left(value: Option<f32>) {
     unsafe {
         CAMERA.pos.x -= value.unwrap_or(CAMERA_SENSITIVITY);
     }
 }
 
 #[wasm_bindgen]
-pub fn camera_right(value: Option<f64>) {
+pub fn camera_right(value: Option<f32>) {
     unsafe {
         CAMERA.pos.x += value.unwrap_or(CAMERA_SENSITIVITY);
     }
 }
 
 #[wasm_bindgen]
-pub fn camera_up(value: Option<f64>) {
+pub fn camera_up(value: Option<f32>) {
     unsafe {
         CAMERA.pos.y -= value.unwrap_or(CAMERA_SENSITIVITY);
     }
 }
 
 #[wasm_bindgen]
-pub fn camera_down(value: Option<f64>) {
+pub fn camera_down(value: Option<f32>) {
     unsafe {
         CAMERA.pos.y += value.unwrap_or(CAMERA_SENSITIVITY);
     }
@@ -57,7 +57,7 @@ pub fn camera_down(value: Option<f64>) {
 
 #[allow(static_mut_refs)]
 #[wasm_bindgen]
-pub fn camera_rotate_left(value: Option<f64>) {
+pub fn camera_rotate_left(value: Option<f32>) {
     unsafe {
         CAMERA.facing.add_yaw(-value.unwrap_or(CAMERA_SENSITIVITY));
     }
@@ -65,7 +65,7 @@ pub fn camera_rotate_left(value: Option<f64>) {
 
 #[allow(static_mut_refs)]
 #[wasm_bindgen]
-pub fn camera_rotate_right(value: Option<f64>) {
+pub fn camera_rotate_right(value: Option<f32>) {
     unsafe {
         CAMERA.facing.add_yaw(value.unwrap_or(CAMERA_SENSITIVITY));
     }
@@ -73,7 +73,7 @@ pub fn camera_rotate_right(value: Option<f64>) {
 
 #[allow(static_mut_refs)]
 #[wasm_bindgen]
-pub fn camera_rotate_up(value: Option<f64>) {
+pub fn camera_rotate_up(value: Option<f32>) {
     unsafe {
         CAMERA.facing.add_pitch(-value.unwrap_or(CAMERA_SENSITIVITY));
     }
@@ -81,7 +81,7 @@ pub fn camera_rotate_up(value: Option<f64>) {
 
 #[allow(static_mut_refs)]
 #[wasm_bindgen]
-pub fn camera_rotate_down(value: Option<f64>) {
+pub fn camera_rotate_down(value: Option<f32>) {
     unsafe {
         CAMERA.facing.add_pitch(value.unwrap_or(CAMERA_SENSITIVITY));
     }
@@ -89,7 +89,7 @@ pub fn camera_rotate_down(value: Option<f64>) {
 
 #[allow(static_mut_refs)]
 #[wasm_bindgen]
-pub fn camera_rotate_anticlockwise(value: Option<f64>) {
+pub fn camera_rotate_anticlockwise(value: Option<f32>) {
     unsafe {
         CAMERA.facing.add_roll(-value.unwrap_or(CAMERA_SENSITIVITY));
     }
@@ -97,7 +97,7 @@ pub fn camera_rotate_anticlockwise(value: Option<f64>) {
 
 #[allow(static_mut_refs)]
 #[wasm_bindgen]
-pub fn camera_rotate_clockwise(value: Option<f64>) {
+pub fn camera_rotate_clockwise(value: Option<f32>) {
     unsafe {
         CAMERA.facing.add_roll(value.unwrap_or(CAMERA_SENSITIVITY));
     }
