+++ wasm/src/positioning/coordinates.rs
@@ -1,7 +1,7 @@
 #[derive(Clone)]
 #[derive(Copy)]
 pub struct Coordinates {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
